backend/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.30.0
alembic>=1.13.1
pgvector>=0.2.5

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Embeddings (local - no API key needed)
sentence-transformers>=3.0.0

# LLM (OpenRouter uses OpenAI-compatible API)
openai>=1.30.0

# Document Parsing
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2

# Text Processing
tiktoken>=0.8.0

# HTTP Client
httpx>=0.27.0

# Telegram
python-telegram-bot>=21.0.1

# Config
pydantic-settings>=2.1.0

# CORS & Utils
python-dotenv>=1.0.1
aiofiles>=23.2.1
